+++ tests/basic_store.py
@@ -10,6 +10,7 @@ from simplekv._compat import BytesIO, xrange, text_type
 from simplekv.decorator import PrefixDecorator
 from simplekv.crypt import HMACDecorator
 from simplekv.idgen import UUIDDecorator, HashDecorator
+from simplekv import CopyMixin
 
 
 class BasicStore(object):
@@ -40,39 +41,26 @@ class BasicStore(object):
         store.put_file(key, BytesIO(value))
         assert store.open(key).read() == value
 
-    def test_store_and_move(self, copy_move_store, key, key2, value):
-        copy_move_store.put(key, value)
-        assert copy_move_store.get(key) == value
-        copy_move_store.move(key, key2)
-        assert key not in copy_move_store
-        assert copy_move_store.get(key2) == value
+    def test_store_and_copy(self, store, key, key2, value):
+        if not isinstance(store, CopyMixin):
+            pytest.skip()
+        store.put(key, value)
+        assert store.get(key) == value
+        store.copy(key, key2)
+        assert store.get(key) == value
+        assert store.get(key2) == value
 
-    def test_store_and_copy(self, copy_move_store, key, key2, value):
-        copy_move_store.put(key, value)
-        assert copy_move_store.get(key) == value
-        copy_move_store.copy(key, key2)
-        assert copy_move_store.get(key) == value
-        assert copy_move_store.get(key2) == value
-
-    def test_store_and_move_overwite(self, copy_move_store, key, key2,
-                                     value, value2):
-        copy_move_store.put(key, value)
-        copy_move_store.put(key2, value2)
-        assert copy_move_store.get(key) == value
-        assert copy_move_store.get(key2) == value2
-        copy_move_store.move(key, key2)
-        assert key not in copy_move_store
-        assert copy_move_store.get(key2) == value
-
-    def test_store_and_copy_overwrite(self, copy_move_store, key, key2,
+    def test_store_and_copy_overwrite(self, store, key, key2,
                                       value, value2):
-        copy_move_store.put(key, value)
-        copy_move_store.put(key2, value2)
-        assert copy_move_store.get(key) == value
-        assert copy_move_store.get(key2) == value2
-        copy_move_store.copy(key, key2)
-        assert copy_move_store.get(key) == value
-        assert copy_move_store.get(key2) == value
+        if not isinstance(store, CopyMixin):
+            pytest.skip()
+        store.put(key, value)
+        store.put(key2, value2)
+        assert store.get(key) == value
+        assert store.get(key2) == value2
+        store.copy(key, key2)
+        assert store.get(key) == value
+        assert store.get(key2) == value
 
     def test_open_incremental_read(self, store, key, long_value):
         store.put_file(key, BytesIO(long_value))