+++ tests/test_dict_store.py
@@ -1,8 +1,6 @@
 #!/usr/bin/env python
 # coding=utf8
 from simplekv.memory import DictStore
-from simplekv import CopyMoveMixin
-
 from basic_store import BasicStore
 from idgens import UUIDGen, HashGen
 from test_hmac import HMACDec
@@ -14,9 +12,3 @@ class TestDictStore(BasicStore, UUIDGen, HashGen, HMACDec):
     @pytest.fixture
     def store(self):
         return DictStore()
-
-    @pytest.fixture()
-    def copy_move_store(self):
-        class CopyMoveStore(DictStore, CopyMoveMixin):
-            pass
-        return CopyMoveStore()