+++ tests/test_mongo.py
@@ -2,7 +2,6 @@
 # coding=utf8
 
 from uuid import uuid4 as uuid
-from simplekv import CopyMoveMixin
 
 import pytest
 pymongo = pytest.importorskip('pymongo')
@@ -23,15 +22,4 @@ class TestMongoDB(BasicStore):
         except pymongo.errors.ConnectionFailure:
             pytest.skip('could not connect to mongodb')
         yield MongoStore(conn[db_name], 'simplekv-tests')
-        conn.drop_database(db_name)
-
-    @pytest.fixture()
-    def copy_move_store(self, db_name):
-        class CopyMoveStore(MongoStore, CopyMoveMixin):
-            pass
-        try:
-            conn = pymongo.MongoClient()
-        except pymongo.errors.ConnectionFailure:
-            pytest.skip('could not connect to mongodb')
-        yield CopyMoveStore(conn[db_name], 'simplekv-tests')
         conn.drop_database(db_name)