+++ tests/test_redis.py
@@ -1,7 +1,6 @@
 #!/usr/bin/env python
 
 from basic_store import BasicStore, TTLStore
-from simplekv import CopyMoveMixin
 
 import pytest
 redis = pytest.importorskip('redis')
@@ -23,22 +22,4 @@ class TestRedisStore(TTLStore, BasicStore):
 
         r.flushdb()
         yield RedisStore(r)
-        r.flushdb()
-
-    @pytest.yield_fixture()
-    def copy_move_store(self):
-        from simplekv.memory.redisstore import RedisStore
-
-        class CopyMoveStore(RedisStore, CopyMoveMixin):
-            pass
-
-        r = StrictRedis()
-
-        try:
-            r.get('anything')
-        except ConnectionError:
-            pytest.skip('Could not connect to redis server')
-
-        r.flushdb()
-        yield CopyMoveStore(r)
         r.flushdb()